requirements.txt
aiofiles==24.1.0
aiogram==2.25.1
aiohappyeyeballs==2.6.1
aiohttp==3.8.6
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.6.3
asgiref==3.10.0
async-timeout==4.0.3
attrs==25.3.0
autobahn==25.10.2
Automat==25.4.16
Babel==2.10.3
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.3.0
cachetools==4.2.2
certifi==2025.4.26
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.2
crispy-bootstrap5==2025.6
cryptography==3.4.8
cycler==0.12.1
daphne==4.2.1
decorator==5.1.1
distlib==0.3.9
Django==6.0
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-filter==25.1
django-otp==1.6.1
django-redis==6.0.0
django-unfold==0.64.2
djangodoo==0.2.3
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
docutils==0.20.1
drf-extra-fields==3.7.0
ebaysdk==2.1.5
ERPpeek==1.7.1
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
fonttools==4.58.1
freezegun==1.2.1
frozenlist==1.7.0
geoip2==2.9.0
googletrans==4.0.0rc1
greenlet==3.2.3
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==24.7.2
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.2
kiwisolver==1.4.8
libsass==0.22.0
lxml==4.9.3
magic-filter==1.0.12
MarkupSafe==2.1.2
matplotlib==3.10.3
maxminddb==3.0.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.6.3
num2words==0.5.10
numpy==2.2.6
ofxparse==0.21
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
passlib==1.7.4
phonenumbers==9.0.10
Pillow==10.1.0
pipenv==2025.0.4
platformdirs==4.3.8
polib==1.1.1
propcache==0.3.2
psutil==5.9.4
psycopg2==2.9.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==1.4.2
pyfiglet==1.0.2
pygame==2.6.1
PyJWT==2.10.1
pymunk==7.1.0
pyOpenSSL==21.0.0
pyotp==2.9.0
pyparsing==3.2.3
pypdf==5.6.0
PyPDF2==2.12.1
pypiwin32==223
pypng==0.20220715.0
pyproject_hooks==1.2.0
pyserial==3.5
pyTelegramBotAPI==4.27.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.1.1
python-memcached==1.62
python-stdnum==1.19
python-telegram-bot==22.3
pytz==2025.2
pyusb==1.2.1
pywin32==311
PyYAML==6.0.3
qrcode==7.4.2
redis==6.4.0
reportlab==3.6.12
requests==2.31.0
requests-file==3.0.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rjsmin==1.2.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.2
sympy==1.14.0
telegram==0.0.1
tornado==6.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.5
uuid==1.30
uuid_utils==0.11.0
uvicorn==0.38.0
virtualenv==20.31.2
vobject==0.9.6.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==2.2.2
wikipedia==1.4.0
Wikipedia-API==0.8.1
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
yarl==1.20.1
zeep==4.2.1
zope.interface==8.0.1